feat(finetune): add bf16 mixed-precision via amp_dtype config
Wraps the forward + loss in torch.autocast for both the tokenizer and
predictor training loops, gated by a new Config.amp_dtype field. Setting
"bfloat16" enables bf16 autocast; None (default) keeps the existing FP32
path bit-exact.

bf16 has the same exponent range as FP32, so AdamW master weights need
no scaling and no GradScaler is wired in. Backward, gradient clipping,
and optimizer steps stay outside autocast as recommended.

Inside Kronos.forward the s1 logits are explicitly cast back to FP32
before the in-model torch.multinomial sampling, so the stochastic path
is invariant across precision policies and bf16 training tracks the
FP32 trajectory exactly when seeds and batches are held fixed.

+++ model/kronos.py
@@ -267,7 +267,9 @@ class Kronos(nn.Module, PyTorchModelHubMixin):
         if use_teacher_forcing:
             sibling_embed = self.embedding.emb_s1(s1_targets)
         else:
-            s1_probs = F.softmax(s1_logits.detach(), dim=-1)
+            # Sampling stays in FP32 even under autocast so the stochastic
+            # path is invariant to the surrounding precision policy.
+            s1_probs = F.softmax(s1_logits.detach().float(), dim=-1)
             sample_s1_ids = torch.multinomial(s1_probs.view(-1, self.s1_vocab_size), 1).view(s1_ids.shape)
             sibling_embed = self.embedding.emb_s1(sample_s1_ids)
